requirements.txt
Django==4.0.2
pycoingecko==2.2.0
djangorestframework==3.13.1
django-cors-headers==3.12.0